+++ src/lib/converter/converter-events.ts
@@ -2,6 +2,7 @@ export const ConverterEvents = {
     BEGIN: "begin",
     END: "end",
     CREATE_DECLARATION: "createDeclaration",
+    CREATE_DOCUMENT: "createDocument",
     CREATE_SIGNATURE: "createSignature",
     CREATE_PARAMETER: "createParameter",
     CREATE_TYPE_PARAMETER: "createTypeParameter",